Fix default config handling
Bukkit already loads the default config entries when the config doesn't
exist, so checkIntegrity() doesn't fail on first load (it checks for
null and the only invalid default, token, is "") so the default config
isn't saved.

This happened to work until recently due to a broken port number check
that has now been fixed.

It turns out though that saveDefaultConfig() already won't replace an
existing config file, so just call that regardless to ensure we get a
config file for the user to edit.

checkIntegrity() is still helpful to warn that the (probably default)
config isn't complete though, so repurpose it for that, adjusting the
error message, dropping the genConfig() code, and fixing
checkIntegrity() to check the strings aren't empty.