WebClient: Fix broken error handling & JSON detection
stream.toString() simply describes the stream without reading anything
from it, so it can't be used by doRequest() to determine whether the
stream will have JSON data (e.g. post chat doesn't). the InputStream
mark() and reset() also don't seem to be supported in this case (for me)
so can't be used.

To preserve JSON exceptions in the case it is wanted, simply add a new
boolean paramter expectJSON and only attempt to read the JSON if true.

This fixes vibeCheck() (returns JSON) without reintroducing the noisy
JSON exceptions on message sends (no JSON).

The error hanlding in doRequest() is fixed in the process to only call
getErrorStream() after getResponseCode() has waited to see if an error
has occurred. Now we call either getInputStream() or getErrorStream()
depending on the status, and in the event of an error we always attempt
to read JSON and always return null (since doRequest() is printing the
actual error message provided in the JSON so the caller probably doesn't
need to know specifics).