Optionally send player join/quit/kick events
Listen for Bukkit player join, quit & kick events, and send them to the
matrix-appservice-minecraft endpoints (along with the kick message) if
configured to do so with relay-minecraft-membership in the config.yml.

This allows matrix-appservice-minecraft to keep the fake matrix user
room membership in sync with the players who are online.