Optionally send player join/quit/kick events
Listen for Bukkit player join, quit & kick events, and send them to the
matrix-appservice-minecraft endpoints along with the join/quit/kick
messages if configured to do so with relay-mc-membership in the
config.yml.

This allows matrix-appservice-minecraft to keep the fake matrix user
room membership in sync with the players who are online.