Optionally handle kicks, bans & unbans from Matrix
If matrix-appservice-minecraft reports kicks, bans or unbans, optionally
relay those on to the Minecraft server if the corresponding
relay-matrix-kicks or relay-matrix-bans are set in the config.yml.

Handling is implemented only for Bukkit at the moment, where we do
currently broadcast the events to players like /kick, /ban & /unban do,
but not with localized text or necessarily matching default reasons
(e.g. EssentialsX reimplements them).

The following Bukkit permissions are added for receiving these
broadcasts (default to op):
 - matrix.kick.notify
 - matrix.ban.notify
 - matrix.unban.notify