MCMessage: Use JSONObject to generate JSON
Refactor MCMessage JSON generation to use JSONObject and delegate the
Player JSON generation to the PoloPlayer class. This will make it
simpler to add new event types other than messages that we can generate
JSON for.

A new MCEvent acts as a base class for events (of which MCMessage is now
one). Child classes must implement toJSON() to generate a JSONObject,
and the base class implements toString().