Bukkit: Pass player's texture (skin) with events
Pass the player's base64 encoded texture (which includes the skin URL) to the
appservice along with events. This avoids the need for the appservice to look
up the texture if the server already has it, and also allows offline mode
players to have their skins show up on the Matrix side.

To get the texture we borrow some helper code from the DynMap plugin, which has
been simplified to just retrieve the texture rather than the skin URL in
particular.

This is unimplemented for sponge.