MCListener: Abstract creation of new PoloPlayer
The PoloPlayer type is created on multiple occasions in the same way. To
simplify the addition of textures and displayName to PoloPlayer,
abstract the creation of a new PoloPlayer into a newPoloPlayer()
function. Also to allow for multiple listener classes when
soft-dependencies are involved, split out potentially common code into a
PoloListener base class.