Bukkit: Handle EssentialsX NickChangeEvent
Handle the NickChangeEvent from the EssentialsX plugin, and respond by
reposting the join event with the new displayName. We wait a tick to
allow the new displayName to be set first, otherwise the new nick needs
to be specially handled.

This allows the displayName to be changed on the matrix side as soon as
the nick is changed on the Minecraft/EssentialsX side.

In order to safely register the listener, we add Essentials as a soft
dependency and register the Essentials EventHandlers in a new
Ess3Listener class.