Bukkit: Send formatted text to players
Use the new common IntermediateJSON class to convert the intermediate
JSON formatted message body provided by the appservice to text
components and send to players with the help of
adventure-platform-bukkit.

We support highlighted variants sent to mentioned players (or all
players if a legitimate @room mention is made), and non-interactive
formatted console output.

Unfortunately due to a bug[1] in adventure-platform the console output
must be deserialized to legacy formatting codes and go through the
logger for now.

[1] https://github.com/KyoriPowered/adventure-platform/issues/38